linked-list/92-reverse-linked-list/reverselinkedlist.go
package main

/**
 * Definition for singly-linked list.
 * type ListNode struct {
 *     Val int
 *     Next *ListNode
 * }
 */

type ListNode struct {
	Val  int
	Next *ListNode
}

func reverseBetween(head *ListNode, left int, right int) *ListNode {
	// 构建虚拟头结点, 避免复杂的边界判断
	// pre: left所在的前一个结点
	// cur: left所在的结点
	dummy := &ListNode{Next: head}
	pre := dummy
	for i := 1; i < left; i++ {
		pre = pre.Next
	}
	cur := pre.Next

	// 反转链表, 从left到right
	// 迭代更新pre.Next和cur.Next
	for i := left; i < right; i++ {
		next := cur.Next
		cur.Next = next.Next
		next.Next = pre.Next
		pre.Next = next
	}

	return dummy.Next
}
